Refactor soil forecasting and health scoring logic
- Removed unnecessary docstrings and comments from various modules to improve code clarity.
- Updated parameter names from `farm_id` to `farmer_id` across multiple schemas and endpoints for consistency.
- Introduced a new `SoilForecastService` class to encapsulate business logic related to soil forecasting, including health score calculations, realignment strategies, and recommendation generation.
- Enhanced the health score calculation method to utilize optimal ranges for rice cultivation.
- Implemented a fallback forecasting mechanism when the ML service is unavailable, using simple rule-based projections.
- Improved the formatting of weekly forecasts and added detailed logging for health score calculations and deviations.
